Added UserProfile to views, reated component for uploading photos in src fixed
